feat: Style SearchBox in BoardViewPage
