+++ application/millage-client/src/components/boards/BoardView/index.ts
@@ -1,2 +1,2 @@
-export {default as BoardBoxTop} from './BoardBoxTop';
+export {default as SearchBox} from './SearchBox';
 export {default as PostListBox} from './PostListBox';
